# Conflicts:
#	src/main/kotlin/fr/insalyon/pld/agile/service/roundmodifier/implementation/RoundModifierImp.kt
update addDelivery + tests

+++ src/main/kotlin/fr/insalyon/pld/agile/model/Junction.kt
@@ -6,7 +6,7 @@ import fr.insalyon.pld.agile.lib.graph.model.Measurable
  * A part of a road between two intersections
  */
 class Junction(
-    override val length: Long,
+    override val length: Int,
     val name: String
 ) : Measurable {
